explicit_{push,pop} quirk + removed out_of_order + cpp files
- removed out_of_order in mpl in favor of get_reference_tuple
- removed unnecessary base + tuple<> functions
- argparse.cpp
- migration to utility.cpp
- header cleanup
